+++ src/App.vue
@@ -10,8 +10,8 @@
 
 <script lang="ts">
 import Vue from 'vue';
-import Navbar from './components/Navbar.vue';
-import Footer from './components/Footer.vue';
+import Navbar from '@/components/layout/TheNavbar.vue';
+import Footer from '@/components/layout/TheFooter.vue';
 import { Component } from 'vue-property-decorator';
 
 @Component({
