+++ src/components/tournament/TournamentGallery.vue
@@ -2,13 +2,12 @@
   <div>
     <v-container>
       <v-row justify="center">
-        
         <gallery
           :images="images"
           :index="index"
           @close="index = null"
         ></gallery>
-        
+
         <v-img
           class="image rounded ma-2"
           v-for="(image, imageIndex) in images"
@@ -20,7 +19,6 @@
             height: '300px',
           }"
         ></v-img>
-
       </v-row>
     </v-container>
   </div>
@@ -30,26 +28,25 @@
 import Vue from 'vue';
 import { Component, Prop } from 'vue-property-decorator';
 import axios from '@/axios';
-import VueGallery from '../../node_modules/vue-gallery/src/component/gallery.vue';
+import VueGallery from '../../../node_modules/vue-gallery/src/component/gallery.vue';
 
 @Component({
   components: {
-    'gallery': VueGallery,
-  }
+    gallery: VueGallery,
+  },
 })
 export default class Gallery extends Vue {
-  @Prop({required: true}) readonly value!: String[]
+  @Prop({ required: true }) readonly value!: String[];
   data() {
     return {
       index: null,
     };
   }
-  
+
   get images() {
     return this.value;
   }
 }
-
 </script>
 
 <style scoped>
@@ -59,8 +56,8 @@ export default class Gallery extends Vue {
   background-position: center center;
 }
 
-@media screen and (min-width: 1040px) and (max-width: 1510px){
-  .image{
+@media screen and (min-width: 1040px) and (max-width: 1510px) {
+  .image {
     max-width: 50%;
   }
 }
