+++ src/components/tournament/TournamentInfo.vue
