+++ src/components/tournament/TournamentPlayers.vue
